Move regression test into separate test class
Since we're using regular `HazelcastInstance` objects and not test variants, our test needs to be executed in a Serial environment; rather than slow down all the other tests in MembershipUpdateTest, let's move it out of the class.
